# Back-end/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Seguros TM",
        "description": "API para gestão de seguros, clientes, apólices e sinistros",
        "contact": {
            "email": "suporte@segurostm.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Servidor de Desenvolvimento"
        }
    ],
    "paths": {
...
        "/api/seguradora/seguros": {
            "get": {
                "tags": [
                    "Seguradora - Seguros"
                ],
                "summary": "Listar seguros da seguradora",
                "description": "Retorna todos os seguros cadastrados pela seguradora autenticada com opções de filtro",
                "operationId": "d1fd771c38335bd89394376410e0a84a",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Filtrar por status (ativo, inativo)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "ativo",
                                "inativo"
                            ]
                        }
                    },
                    {
                        "name": "tipo_seguro",
                        "in": "query",
                        "description": "Filtrar por tipo de seguro (automovel, saude, residencial, vida)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "automovel",
                                "saude",
                                "residencial",
                                "vida"
                            ]
                        }
                    },
                    {
                        "name": "id_categoria",
                        "in": "query",
                        "description": "Filtrar por ID da categoria",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Número de registros por página",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 15
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de seguros retornada com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "current_page": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id_seguro": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "nome": {
                                                        "type": "string",
                                                        "example": "Seguro Auto Premium"
                                                    },
                                                    "descricao": {
                                                        "type": "string",
                                                        "example": "Cobertura completa para veículos"
                                                    },
                                                    "tipo_seguro": {
                                                        "type": "string",
                                                        "example": "automovel"
                                                    },
                                                    "status": {
                                                        "type": "string",
                                                        "example": "ativo"
                                                    },
                                                    "id_categoria": {
                                                        "type": "integer",
                                                        "example": 1
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "total": {
                                            "type": "integer",
                                            "example": 50
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autenticado"
                    },
                    "500": {
                        "description": "Erro ao listar seguros"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Seguradora - Seguros"
                ],
                "summary": "Criar novo seguro",
                "description": "Cadastra um novo produto de seguro para a seguradora autenticada",
                "operationId": "a7940fa1bb43a7c6026e532d8f54f54f",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nome",
                                    "tipo_seguro",
                                    "id_categoria"
                                ],
                                "properties": {
                                    "nome": {
                                        "description": "Nome do seguro",
                                        "type": "string",
                                        "example": "Seguro Auto Premium"
                                    },
                                    "descricao": {
                                        "description": "Descrição detalhada",
                                        "type": "string",
                                        "example": "Cobertura completa para veículos de passeio"
                                    },
                                    "tipo_seguro": {
                                        "description": "Tipo de seguro",
                                        "type": "string",
                                        "enum": [
                                            "automovel",
                                            "saude",
                                            "residencial",
                                            "vida"
                                        ],
                                        "example": "automovel"
                                    },
                                    "id_categoria": {
                                        "description": "ID da categoria do seguro",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "detalhes": {
                                        "description": "Informações adicionais específicas do tipo de seguro",
                                        "type": "object"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Seguro criado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Seguro criado com sucesso"
                                        },
                                        "data": {
                                            "properties": {
                                                "id_seguro": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "nome": {
                                                    "type": "string",
                                                    "example": "Seguro Auto Premium"
                                                },
                                                "status": {
                                                    "type": "string",
                                                    "example": "ativo"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Erro de validação"
                    },
                    "401": {
                        "description": "Não autenticado"
                    },
                    "500": {
                        "description": "Erro ao criar seguro"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/seguradora/seguros/{id}": {
            "get": {
                "tags": [
                    "Seguradora - Seguros"
                ],
                "summary": "Obter detalhes de um seguro",
                "description": "Retorna informações detalhadas de um seguro específico incluindo preços e coberturas",
                "operationId": "ff98132ac42c8d9da142912f0b62aff1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do seguro",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalhes do seguro",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "id_seguro": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "nome": {
                                                    "type": "string",
                                                    "example": "Seguro Auto Premium"
                                                },
                                                "descricao": {
                                                    "type": "string"
                                                },
                                                "tipo_seguro": {
                                                    "type": "string",
                                                    "example": "automovel"
                                                },
                                                "status": {
                                                    "type": "string",
                                                    "example": "ativo"
                                                },
                                                "precos": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "object"
                                                    }
                                                },
                                                "coberturas": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "object"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Seguro não encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/seguradora/seguros/{id}/precos": {
            "post": {
                "tags": [
                    "Seguradora - Seguros"
                ],
                "summary": "Adicionar preço ao seguro",
                "description": "Adiciona uma nova tabela de preços para um seguro",
                "operationId": "a6b4d08cad4e683f4fa399a53a1fbf10",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do seguro",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "valor_base",
                                    "periodicidade"
                                ],
                                "properties": {
                                    "valor_base": {
                                        "description": "Valor base do prêmio",
                                        "type": "number",
                                        "format": "float",
                                        "example": 1500
                                    },
                                    "periodicidade": {
                                        "type": "string",
                                        "enum": [
                                            "mensal",
                                            "trimestral",
                                            "semestral",
                                            "anual"
                                        ],
                                        "example": "mensal"
                                    },
                                    "descricao": {
                                        "type": "string",
                                        "example": "Plano básico mensal"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Preço adicionado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Preço adicionado com sucesso"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Autenticação JWT. Use o token retornado no endpoint de login.",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}